# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API Documentation",
        "contact": {
            "email": "info@yeagger.com"
        },
        "version": "0.1"
    },
    "servers": [
        {
            "url": "https://foo.localhost:8000/api/",
            "description": "Learning env"
        }
    ],
    "paths": {
...
        "/profiles/{id}": {
            "get": {
                "tags": [
                    "Profiles"
                ],
                "summary": "Get single profile",
                "description": "Returns a single profile based on ID",
                "operationId": "getSingleProfile",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Profile id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
    },
    "security": [
        []
    ]
}